front: split and improve stdcm page modal for e2e tests
