+++ front/tests/006-stdcm.spec.ts
@@ -1,9 +1,14 @@
 import type { Infra, TowedRollingStock } from 'common/api/osrdEditoastApi';
 
-import { electricRollingStockName, fastRollingStockName } from './assets/project-const';
+import { electricRollingStockName, fastRollingStockName } from './assets/constants/project-const';
 import test from './logging-fixture';
-import STDCMLinkedTrainPage from './pages/stdcm-linked-train-page-model';
-import STDCMPage from './pages/stdcm-page-model';
+import ConsistSection from './pages/stdcm/consist-section';
+import DestinationSection from './pages/stdcm/destination-section';
+import LinkedTrainSection from './pages/stdcm/linked-train-section';
+import OriginSection from './pages/stdcm/origin-section';
+import SimulationResultPage from './pages/stdcm/simulation-results-page';
+import STDCMPage from './pages/stdcm/stdcm-page';
+import ViaSection from './pages/stdcm/via-section';
 import { handleAndVerifyInput, waitForInfraStateToBeCached } from './utils';
 import { getInfra, setTowedRollingStock } from './utils/api-setup';
 import type { ConsistFields } from './utils/types';
@@ -13,7 +18,12 @@ test.describe('Verify stdcm simulation page', () => {
   test.use({ viewport: { width: 1920, height: 1080 } });
 
   let stdcmPage: STDCMPage;
-  let stdcmLinkedTrainPage: STDCMLinkedTrainPage;
+  let consistSection: ConsistSection;
+  let originSection: OriginSection;
+  let viaSection: ViaSection;
+  let destinationSection: DestinationSection;
+  let linkedTrainSection: LinkedTrainSection;
+  let simulationResultPage: SimulationResultPage;
 
   let infra: Infra;
   let createdTowedRollingStock: TowedRollingStock;
@@ -44,7 +54,24 @@ test.describe('Verify stdcm simulation page', () => {
   });
 
   test.beforeEach('Navigate to the STDCM page', async ({ page }) => {
-    [stdcmPage, stdcmLinkedTrainPage] = [new STDCMPage(page), new STDCMLinkedTrainPage(page)];
+    [
+      stdcmPage,
+      consistSection,
+      originSection,
+      viaSection,
+      destinationSection,
+      simulationResultPage,
+      linkedTrainSection,
+    ] = [
+      new STDCMPage(page),
+      new ConsistSection(page),
+      new OriginSection(page),
+      new ViaSection(page),
+      new DestinationSection(page),
+      new SimulationResultPage(page),
+      new LinkedTrainSection(page),
+    ];
+
     await page.goto('/stdcm');
     await page.waitForLoadState('networkidle');
     await stdcmPage.removeViteOverlay();
@@ -58,21 +85,23 @@ test.describe('Verify stdcm simulation page', () => {
     // Verify visibility of STDCM elements and handle default fields
 
     await stdcmPage.verifyStdcmElementsVisibility();
-    await stdcmPage.verifyAllDefaultPageFields();
-    await stdcmPage.addAndDeletedDefaultVia();
-    await stdcmLinkedTrainPage.addAndDeleteDefaultLinkedPath();
+    await consistSection.verifyDefaultConsistFields();
+    await originSection.verifyDefaultOriginFields();
+    await destinationSection.verifyDefaultDestinationFields();
+    await viaSection.addAndDeletedDefaultVia();
+    await linkedTrainSection.addAndDeleteDefaultLinkedPath();
   });
 
   /** *************** Test 2 **************** */
   test('Launch STDCM simulation with all stops', async () => {
     // Populate STDCM page with origin, destination, and via details, then verify
-    await stdcmPage.fillAndVerifyConsistDetails(
+    await consistSection.fillAndVerifyConsistDetails(
       consistDetails,
       tractionEnginePrefilledValues.tonnage,
       tractionEnginePrefilledValues.length
     );
-    await stdcmPage.fillAndVerifyOriginDetails();
-    await stdcmPage.fillAndVerifyDestinationDetails();
+    await originSection.fillAndVerifyOriginDetails();
+    await destinationSection.fillAndVerifyDestinationDetails();
     const viaDetails = [
       { viaNumber: 1, ciSearchText: 'mid_west' },
       { viaNumber: 2, ciSearchText: 'mid_east' },
@@ -80,7 +109,7 @@ test.describe('Verify stdcm simulation page', () => {
     ];
 
     for (const viaDetail of viaDetails) {
-      await stdcmPage.fillAndVerifyViaDetails(viaDetail);
+      await viaSection.fillAndVerifyViaDetails(viaDetail);
     }
     // Launch simulation and verify output data matches expected results
     await stdcmPage.launchSimulation();
@@ -94,28 +123,28 @@ test.describe('Verify stdcm simulation page', () => {
       towedRollingStock: createdTowedRollingStock.name,
     };
 
-    await stdcmPage.fillAndVerifyConsistDetails(
+    await consistSection.fillAndVerifyConsistDetails(
       towedConsistDetails,
       fastRollingStockPrefilledValues.tonnage,
       fastRollingStockPrefilledValues.length,
       towedRollingStockPrefilledValues.tonnage,
       towedRollingStockPrefilledValues.length
     );
-    await stdcmPage.fillOriginDetailsLight();
-    await stdcmPage.fillDestinationDetailsLight();
-    await stdcmPage.fillAndVerifyViaDetails({
+    await originSection.fillOriginDetailsLight();
+    await destinationSection.fillDestinationDetailsLight();
+    await viaSection.fillAndVerifyViaDetails({
       viaNumber: 1,
       ciSearchText: 'mid_west',
     });
     // Run first simulation without capacity
     await stdcmPage.launchSimulation();
-    await stdcmPage.verifySimulationDetails({
+    await simulationResultPage.verifySimulationDetails({
       simulationNumber: 1,
     });
     // Update tonnage and launch a second simulation with capacity
-    await handleAndVerifyInput(stdcmPage.dateOriginArrival, UPDATED_ORIGIN_ARRIVAL_DATE);
+    await handleAndVerifyInput(originSection.dateOriginArrival, UPDATED_ORIGIN_ARRIVAL_DATE);
     await stdcmPage.launchSimulation();
-    await stdcmPage.verifySimulationDetails({
+    await simulationResultPage.verifySimulationDetails({
       simulationNumber: 2,
       simulationLengthAndDuration: '51 km — 2h 35min',
     });
